Redid how EPUB import is done
Use IDs in categories.yaml to ensure that the filenames are preserved
on import, so it’s easy to offer download links to the user later. We
could have just linked to the files but as the names change by date
with each new build, it is better to use IDs and just give the import
script a directory to search through for each one.